--- src/applications/check-in/reducer/index.js
@@ -1,9 +1,0 @@
-const initialState = {};
-
-const checkInReducer = (state = initialState, _action) => {
-  return { ...state };
-};
-
-export default {
-  checkInData: checkInReducer,
-};
