+++ src/applications/check-in/reducers/index.js
@@ -1,1 +1,9 @@
-export default {};
+const initialState = {};
+
+const checkInReducer = (state = initialState, _action) => {
+  return { ...state };
+};
+
+export default {
+  checkInData: checkInReducer,
+};
